Bot name changed to tra4me.
tra4me name is inspired in the phrase "Trade for me".

The name of the bot was changed for the following
reasons:

1: tra4me is an intuitive name that describes the
purpose of this software.

2: There are other bots named "tbot", but so far I
have not found a bot named tra4me.

3: A Github project will be created, and so far there
are no projects in Github named tra4me.

I did this name change on this feature branch because it
is the branch that has the most stable code and files.